+++ niryo_robot_python_ros_wrapper/src/niryo_robot_python_ros_wrapper/ros_wrapper.py
@@ -729,7 +729,7 @@ class NiryoRosWrapper(AbstractNiryoRosWrapper):
         :return: list of joints value
         :rtype: list[float]
         """
-        result = self._call_service('/niryo_robot_commander/stop_command', Trigger)
+        result = self._call_service('/niryo_robot_arm_commander/stop_command', Trigger)
         return self._classic_return_w_check(result)
 
     def set_jog_use_state(self, state):